+++ shotmanager/overlay_tools/interact_shots_stack/shots_stack_toolbar.py
@@ -69,6 +69,23 @@ def draw_shots_stack_toolbar_in_editor(self, context):
 
     row.separator(factor=0.8)
 
+    if config.devDebug:
+        # row.operator(
+        #     "uas_gpu.dopesheet_gpu_sample",
+        #     text="GPU",
+        #     depress=context.window_manager.UAS_shot_manager_display_dopesheet_gpu_sample,
+        #     # icon_value=icon.icon_id,
+        # )
+        row.prop(
+            context.window_manager,
+            "UAS_shot_manager_display_dopesheet_gpu_sample",
+            text="GPU",
+            # depress=context.window_manager.UAS_shot_manager_display_dopesheet_gpu_sample,
+            # icon_value=icon.icon_id,
+        )
+
+    row.separator(factor=0.8)
+
     row.operator(
         "uas_shot_manager.toggle_shots_stack_with_overlay_tools",
         text="",